requirements.txt
discord.py==2.3.2
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0
python-dotenv==1.0.1